Make horizonclient examples fully qualified (#1867)
* horizonclient: Make asset_request examples fully qualified

* horizonclient: Make effect_request examples fully qualified

* horizonclient: Make ledger_request examples fully qualified

* horizonclient: Make offer_request examples fully qualified

* horizonclient: Make operation_request examples fully qualified

* horizonclient: Make order_book_request examples fully qualified

* horizonclient: Make paths_request examples fully qualified

* horizonclient: Make root example fully qualified

* horizonclient: Make trade_aggregation examples fully qualified

* horizonclient: Make trade_request examples fully qualified

* horizonclient: Make transaction_request examples fully qualified

* horizonclient: Make main examples fully qualified

* horizonclient: Remove empty lines and reformat examples_test.go
