+++ clients/horizonclient/asset_request_test.go
@@ -1,7 +1,6 @@
 package horizonclient
 
 import (
-	"fmt"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -29,62 +28,4 @@ func TestAssetRequestBuildUrl(t *testing.T) {
 	// It should return valid assets endpoint and no errors
 	require.NoError(t, err)
 	assert.Equal(t, "assets?asset_code=ABC&asset_issuer=GCLWGQPMKXQSPF776IU33AH4PZNOOWNAWGGKVTBQMIC5IMKUNP3E6NVU&order=desc", endpoint)
-}
-
-func ExampleClient_NextAssetsPage() {
-	client := DefaultPublicNetClient
-	// assets for asset issuer
-	assetRequest := AssetRequest{ForAssetIssuer: "GCLWGQPMKXQSPF776IU33AH4PZNOOWNAWGGKVTBQMIC5IMKUNP3E6NVU",
-		Limit: 20}
-	asset, err := client.Assets(assetRequest)
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-	fmt.Print(asset)
-
-	// all assets
-	assetRequest = AssetRequest{}
-	asset, err = client.Assets(assetRequest)
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-
-	// next page
-	nextPage, err := client.NextAssetsPage(asset)
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-	fmt.Println(nextPage)
-}
-
-func ExampleClient_PrevAssetsPage() {
-	client := DefaultPublicNetClient
-	// assets for asset issuer
-	assetRequest := AssetRequest{ForAssetIssuer: "GCLWGQPMKXQSPF776IU33AH4PZNOOWNAWGGKVTBQMIC5IMKUNP3E6NVU",
-		Limit: 20}
-	asset, err := client.Assets(assetRequest)
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-	fmt.Print(asset)
-
-	// all assets
-	assetRequest = AssetRequest{}
-	asset, err = client.Assets(assetRequest)
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-
-	// next page
-	prevPage, err := client.PrevAssetsPage(asset)
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-	fmt.Println(prevPage)
 }
