+++ clients/horizonclient/main_test.go
@@ -1,7 +1,6 @@
 package horizonclient
 
 import (
-	"fmt"
 	"net/http"
 	"testing"
 	"time"
@@ -17,292 +16,6 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func ExampleClient_AccountDetail() {
-
-	client := DefaultPublicNetClient
-	accountRequest := AccountRequest{AccountID: "GCLWGQPMKXQSPF776IU33AH4PZNOOWNAWGGKVTBQMIC5IMKUNP3E6NVU"}
-
-	account, err := client.AccountDetail(accountRequest)
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-
-	fmt.Print(account)
-}
-
-func ExampleClient_Effects() {
-
-	client := DefaultPublicNetClient
-	// effects for an account
-	effectRequest := EffectRequest{ForAccount: "GCLWGQPMKXQSPF776IU33AH4PZNOOWNAWGGKVTBQMIC5IMKUNP3E6NVU"}
-	effect, err := client.Effects(effectRequest)
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-	fmt.Print(effect)
-
-	// all effects
-	effectRequest = EffectRequest{}
-	effect, err = client.Effects(effectRequest)
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-	records := effect.Embedded.Records
-	if records[0].GetType() == "account_created" {
-		acc, ok := records[0].(effects.AccountCreated)
-		if ok {
-			fmt.Print(acc.Account)
-			fmt.Print(acc.StartingBalance)
-		}
-	}
-}
-
-func ExampleClient_Assets() {
-
-	client := DefaultPublicNetClient
-	// assets for asset issuer
-	assetRequest := AssetRequest{ForAssetIssuer: "GCLWGQPMKXQSPF776IU33AH4PZNOOWNAWGGKVTBQMIC5IMKUNP3E6NVU"}
-	asset, err := client.Assets(assetRequest)
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-	fmt.Print(asset)
-
-	// all assets
-	assetRequest = AssetRequest{}
-	asset, err = client.Assets(assetRequest)
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-	fmt.Print(asset)
-}
-
-func ExampleClient_LedgerDetail() {
-
-	client := DefaultPublicNetClient
-	// details for a ledger
-	sequence := uint32(12345)
-	ledger, err := client.LedgerDetail(sequence)
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-	fmt.Print(ledger)
-
-}
-
-func ExampleClient_Metrics() {
-
-	client := DefaultPublicNetClient
-	// horizon metrics
-	metrics, err := client.Metrics()
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-	fmt.Print(metrics)
-
-}
-
-func ExampleClient_FeeStats() {
-
-	client := DefaultPublicNetClient
-	// horizon fees
-	fees, err := client.FeeStats()
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-	fmt.Print(fees)
-
-}
-
-func ExampleClient_Offers() {
-
-	client := DefaultPublicNetClient
-	offerRequest := OfferRequest{ForAccount: "GCLWGQPMKXQSPF776IU33AH4PZNOOWNAWGGKVTBQMIC5IMKUNP3E6NVU", Cursor: "now", Order: OrderDesc}
-	offers, err := client.Offers(offerRequest)
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-	fmt.Print(offers)
-}
-
-func ExampleClient_Operations() {
-
-	client := DefaultPublicNetClient
-	// operations for an account
-	opRequest := OperationRequest{ForAccount: "GCLWGQPMKXQSPF776IU33AH4PZNOOWNAWGGKVTBQMIC5IMKUNP3E6NVU"}
-	ops, err := client.Operations(opRequest)
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-	fmt.Print(ops)
-
-	// all operations
-	opRequest = OperationRequest{Cursor: "now"}
-	ops, err = client.Operations(opRequest)
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-	fmt.Print(ops)
-	records := ops.Embedded.Records
-
-	for _, value := range records {
-		// prints the type
-		fmt.Print(value.GetType())
-		// for example if the type is change_trust
-		c, ok := value.(operations.ChangeTrust)
-		if ok {
-			// access ChangeTrust fields
-			fmt.Print(c.Trustee)
-		}
-
-	}
-}
-
-func ExampleClient_OperationDetail() {
-
-	client := DefaultPublicNetClient
-	opID := "123456"
-	// operation details for an id
-	ops, err := client.OperationDetail(opID)
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-	fmt.Print(ops)
-}
-
-func ExampleClient_SubmitTransactionXDR() {
-
-	client := DefaultPublicNetClient
-	// https://www.stellar.org/laboratory/#xdr-viewer?input=AAAAAOoS%2F5V%2BBiCPXRiVcz8YsnkDdODufq%2Bg7xdqTdIXN8vyAAAE4gFiW0YAAALxAAAAAQAAAAAAAAAAAAAAAFyuBUcAAAABAAAABzIyMjgyNDUAAAAAAQAAAAEAAAAALhsY%2FFdAHXllTmb025DtCVBw06WDSQjq6I9NrCQHOV8AAAABAAAAAHT8zKV7bRQzuGTpk9AO3gjWJ9jVxBXTgguFORkxHVIKAAAAAAAAAAAAOnDwAAAAAAAAAAIkBzlfAAAAQPefqlsOvni6xX1g3AqddvOp1GOM88JYzayGZodbzTfV5toyhxZvL1ZggY3prFsvrereugEpj1kyPJ67z6gcRg0XN8vyAAAAQGwmoTssW49gaze8iQkz%2FUA2E2N%2BBOo%2B6v7YdOSsvIcZnMc37KmXH920nLosKpDLqkNChVztSZFcbVUlHhjbQgA%3D&type=TransactionEnvelope&network=public
-	txXdr := `AAAAAOoS/5V+BiCPXRiVcz8YsnkDdODufq+g7xdqTdIXN8vyAAAE4gFiW0YAAALxAAAAAQAAAAAAAAAAAAAAAFyuBUcAAAABAAAABzIyMjgyNDUAAAAAAQAAAAEAAAAALhsY/FdAHXllTmb025DtCVBw06WDSQjq6I9NrCQHOV8AAAABAAAAAHT8zKV7bRQzuGTpk9AO3gjWJ9jVxBXTgguFORkxHVIKAAAAAAAAAAAAOnDwAAAAAAAAAAIkBzlfAAAAQPefqlsOvni6xX1g3AqddvOp1GOM88JYzayGZodbzTfV5toyhxZvL1ZggY3prFsvrereugEpj1kyPJ67z6gcRg0XN8vyAAAAQGwmoTssW49gaze8iQkz/UA2E2N+BOo+6v7YdOSsvIcZnMc37KmXH920nLosKpDLqkNChVztSZFcbVUlHhjbQgA=`
-
-	// submit transaction
-	resp, err := client.SubmitTransactionXDR(txXdr)
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-
-	fmt.Print(resp)
-	// Output: {{{https://horizon.stellar.org/transactions/f8a09e8a17fc828a1b99814818ddc931876eec0fe9c203f5980d26d92641e1c2 false}} f8a09e8a17fc828a1b99814818ddc931876eec0fe9c203f5980d26d92641e1c2 23350654 AAAAAOoS/5V+BiCPXRiVcz8YsnkDdODufq+g7xdqTdIXN8vyAAAE4gFiW0YAAALxAAAAAQAAAAAAAAAAAAAAAFyuBUcAAAABAAAABzIyMjgyNDUAAAAAAQAAAAEAAAAALhsY/FdAHXllTmb025DtCVBw06WDSQjq6I9NrCQHOV8AAAABAAAAAHT8zKV7bRQzuGTpk9AO3gjWJ9jVxBXTgguFORkxHVIKAAAAAAAAAAAAOnDwAAAAAAAAAAIkBzlfAAAAQPefqlsOvni6xX1g3AqddvOp1GOM88JYzayGZodbzTfV5toyhxZvL1ZggY3prFsvrereugEpj1kyPJ67z6gcRg0XN8vyAAAAQGwmoTssW49gaze8iQkz/UA2E2N+BOo+6v7YdOSsvIcZnMc37KmXH920nLosKpDLqkNChVztSZFcbVUlHhjbQgA= AAAAAAAABOIAAAAAAAAAAQAAAAAAAAABAAAAAAAAAAA= AAAAAQAAAAIAAAADAWRNfgAAAAAAAAAA6hL/lX4GII9dGJVzPxiyeQN04O5+r6DvF2pN0hc3y/IAAAAAAuyTvgFiW0YAAALwAAAAAAAAAAAAAAAAAAAAAAEAAAAAAAAAAAAAAAAAAAAAAAABAWRNfgAAAAAAAAAA6hL/lX4GII9dGJVzPxiyeQN04O5+r6DvF2pN0hc3y/IAAAAAAuyTvgFiW0YAAALxAAAAAAAAAAAAAAAAAAAAAAEAAAAAAAAAAAAAAAAAAAAAAAABAAAABAAAAAMBZE0IAAAAAAAAAAB0/Myle20UM7hk6ZPQDt4I1ifY1cQV04ILhTkZMR1SCgAAAbZToYkOAToKfwAAAAEAAAAAAAAAAAAAAAAAAAAAAQAAAAAAAAAAAAAAAAAAAAAAAAEBZE1+AAAAAAAAAAB0/Myle20UM7hk6ZPQDt4I1ifY1cQV04ILhTkZMR1SCgAAAbZT2/n+AToKfwAAAAEAAAAAAAAAAAAAAAAAAAAAAQAAAAAAAAAAAAAAAAAAAAAAAAMBZE19AAAAAAAAAAAuGxj8V0AdeWVOZvTbkO0JUHDTpYNJCOroj02sJAc5XwAAAACrUfjvARGUKgAApRsAAAAAAAAAAAAAAAAAAAAObnlhbmRldi1pZC5vcmcAAAEAAAAAAAAAAAAAAAAAAAAAAAABAWRNfgAAAAAAAAAALhsY/FdAHXllTmb025DtCVBw06WDSQjq6I9NrCQHOV8AAAAAqxeH/wERlCoAAKUbAAAAAAAAAAAAAAAAAAAADm55YW5kZXYtaWQub3JnAAABAAAAAAAAAAAAAAAAAAAA}
-}
-
-func ExampleClient_SetHorizonTimeOut() {
-
-	client := DefaultTestNetClient
-
-	// https://www.stellar.org/laboratory/#xdr-viewer?input=AAAAABB90WssODNIgi6BHveqzxTRmIpvAFRyVNM%2BHm2GVuCcAAAAZAAABD0AAuV%2FAAAAAAAAAAAAAAABAAAAAAAAAAAAAAAAyTBGxOgfSApppsTnb%2FYRr6gOR8WT0LZNrhLh4y3FCgoAAAAXSHboAAAAAAAAAAABhlbgnAAAAEAivKe977CQCxMOKTuj%2BcWTFqc2OOJU8qGr9afrgu2zDmQaX5Q0cNshc3PiBwe0qw%2F%2BD%2FqJk5QqM5dYeSUGeDQP&type=TransactionEnvelope&network=test
-	txXdr := `AAAAABB90WssODNIgi6BHveqzxTRmIpvAFRyVNM+Hm2GVuCcAAAAZAAABD0AAuV/AAAAAAAAAAAAAAABAAAAAAAAAAAAAAAAyTBGxOgfSApppsTnb/YRr6gOR8WT0LZNrhLh4y3FCgoAAAAXSHboAAAAAAAAAAABhlbgnAAAAEAivKe977CQCxMOKTuj+cWTFqc2OOJU8qGr9afrgu2zDmQaX5Q0cNshc3PiBwe0qw/+D/qJk5QqM5dYeSUGeDQP`
-
-	// test user timeout
-	client = client.SetHorizonTimeOut(30)
-	resp, err := client.SubmitTransactionXDR(txXdr)
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-
-	fmt.Print(resp)
-}
-
-func ExampleClient_Transactions() {
-
-	client := DefaultPublicNetClient
-	// transactions for an account
-	txRequest := TransactionRequest{ForAccount: "GCLWGQPMKXQSPF776IU33AH4PZNOOWNAWGGKVTBQMIC5IMKUNP3E6NVU"}
-	txs, err := client.Transactions(txRequest)
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-	fmt.Print(txs)
-
-	// all transactions
-	txRequest = TransactionRequest{Cursor: "now", Order: OrderDesc}
-	txs, err = client.Transactions(txRequest)
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-	fmt.Print(txs)
-	records := txs.Embedded.Records
-
-	for _, tx := range records {
-		fmt.Print(tx)
-	}
-}
-
-func ExampleClient_OrderBook() {
-
-	client := DefaultPublicNetClient
-	// orderbook for an asset pair, e.g XLM/NGN
-	obRequest := OrderBookRequest{BuyingAssetType: AssetTypeNative, SellingAssetCode: "USD", SellingAssetType: AssetType4, SellingAssetIssuer: "GCLWGQPMKXQSPF776IU33AH4PZNOOWNAWGGKVTBQMIC5IMKUNP3E6NVU"}
-	obs, err := client.OrderBook(obRequest)
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-	fmt.Print(obs)
-}
-
-func ExampleClient_Payments() {
-
-	client := DefaultPublicNetClient
-	// payments for an account
-	opRequest := OperationRequest{ForAccount: "GCLWGQPMKXQSPF776IU33AH4PZNOOWNAWGGKVTBQMIC5IMKUNP3E6NVU"}
-	ops, err := client.Payments(opRequest)
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-	fmt.Print(ops)
-
-	// all payments
-	opRequest = OperationRequest{Cursor: "now"}
-	ops, err = client.Payments(opRequest)
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-	fmt.Print(ops)
-	records := ops.Embedded.Records
-
-	for _, value := range records {
-		// prints the type
-		fmt.Print(value.GetType())
-		// for example if the type is create_account
-		c, ok := value.(operations.CreateAccount)
-		if ok {
-			// access create_account fields
-			fmt.Print(c.StartingBalance)
-		}
-
-	}
-}
-
-func ExampleClient_Fund() {
-	client := DefaultTestNetClient
-	// fund an account
-	resp, err := client.Fund("GCLWGQPMKXQSPF776IU33AH4PZNOOWNAWGGKVTBQMIC5IMKUNP3E6NVU")
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-	fmt.Print(resp)
-}
-
 func TestFixHTTP(t *testing.T) {
 	client := &Client{
 		HorizonURL: "https://localhost/",
@@ -312,64 +25,6 @@ func TestFixHTTP(t *testing.T) {
 	client.Root()
 	// When a request is made, default HTTP client is set
 	assert.IsType(t, client.HTTP, &http.Client{})
-}
-
-func TestClientFund(t *testing.T) {
-	hmock := httptest.NewClient()
-	client := &Client{
-		HorizonURL: "https://localhost/",
-		HTTP:       hmock,
-	}
-
-	testAccount := "GCLWGQPMKXQSPF776IU33AH4PZNOOWNAWGGKVTBQMIC5IMKUNP3E6NVU"
-
-	// not testnet
-	hmock.On(
-		"GET",
-		"https://localhost/friendbot?addr=GCLWGQPMKXQSPF776IU33AH4PZNOOWNAWGGKVTBQMIC5IMKUNP3E6NVU",
-	).ReturnString(200, txSuccess)
-
-	_, err := client.Fund(testAccount)
-	// error case: not testnet
-	if assert.Error(t, err) {
-		assert.Contains(t, err.Error(), "can't fund account from friendbot on production network")
-	}
-
-	// happy path
-	hmock.On(
-		"GET",
-		"https://localhost/friendbot?addr=GCLWGQPMKXQSPF776IU33AH4PZNOOWNAWGGKVTBQMIC5IMKUNP3E6NVU",
-	).ReturnString(200, txSuccess)
-
-	client.isTestNet = true
-	resp, err := client.Fund(testAccount)
-
-	if assert.NoError(t, err) {
-		assert.IsType(t, resp, hProtocol.TransactionSuccess{})
-		assert.Equal(t, resp.Links.Transaction.Href, "https://horizon-testnet.stellar.org/transactions/bcc7a97264dca0a51a63f7ea971b5e7458e334489673078bb2a34eb0cce910ca")
-		assert.Equal(t, resp.Hash, "bcc7a97264dca0a51a63f7ea971b5e7458e334489673078bb2a34eb0cce910ca")
-		assert.Equal(t, resp.Ledger, int32(354811))
-		assert.Equal(t, resp.Env, `AAAAABB90WssODNIgi6BHveqzxTRmIpvAFRyVNM+Hm2GVuCcAAAAZAAABD0AAuV/AAAAAAAAAAAAAAABAAAAAAAAAAAAAAAAyTBGxOgfSApppsTnb/YRr6gOR8WT0LZNrhLh4y3FCgoAAAAXSHboAAAAAAAAAAABhlbgnAAAAEAivKe977CQCxMOKTuj+cWTFqc2OOJU8qGr9afrgu2zDmQaX5Q0cNshc3PiBwe0qw/+D/qJk5QqM5dYeSUGeDQP`)
-		assert.Equal(t, resp.Result, "AAAAAAAAAGQAAAAAAAAAAQAAAAAAAAAAAAAAAAAAAAA=")
-		assert.Equal(t, resp.Meta, `AAAAAQAAAAIAAAADAAVp+wAAAAAAAAAAEH3Rayw4M0iCLoEe96rPFNGYim8AVHJU0z4ebYZW4JwACBP/TuycHAAABD0AAuV+AAAAAAAAAAAAAAAAAAAAAAEAAAAAAAAAAAAAAAAAAAAAAAABAAVp+wAAAAAAAAAAEH3Rayw4M0iCLoEe96rPFNGYim8AVHJU0z4ebYZW4JwACBP/TuycHAAABD0AAuV/AAAAAAAAAAAAAAAAAAAAAAEAAAAAAAAAAAAAAAAAAAAAAAABAAAAAwAAAAMABWn7AAAAAAAAAAAQfdFrLDgzSIIugR73qs8U0ZiKbwBUclTTPh5thlbgnAAIE/9O7JwcAAAEPQAC5X8AAAAAAAAAAAAAAAAAAAAAAQAAAAAAAAAAAAAAAAAAAAAAAAEABWn7AAAAAAAAAAAQfdFrLDgzSIIugR73qs8U0ZiKbwBUclTTPh5thlbgnAAIE+gGdbQcAAAEPQAC5X8AAAAAAAAAAAAAAAAAAAAAAQAAAAAAAAAAAAAAAAAAAAAAAAAABWn7AAAAAAAAAADJMEbE6B9ICmmmxOdv9hGvqA5HxZPQtk2uEuHjLcUKCgAAABdIdugAAAVp+wAAAAAAAAAAAAAAAAAAAAAAAAAAAQAAAAAAAAAAAAAAAAAAAA==`)
-	}
-
-	// failure response
-	hmock.On(
-		"GET",
-		"https://localhost/friendbot?addr=GCLWGQPMKXQSPF776IU33AH4PZNOOWNAWGGKVTBQMIC5IMKUNP3E6NVU",
-	).ReturnString(400, transactionFailure)
-
-	_, err = client.Fund(testAccount)
-	if assert.Error(t, err) {
-		assert.Contains(t, err.Error(), "horizon error")
-		horizonError, ok := errors.Cause(err).(*Error)
-		assert.Equal(t, ok, true)
-		assert.Equal(t, horizonError.Problem.Title, "Transaction Failed")
-		resultString, err := horizonError.ResultString()
-		assert.Nil(t, err)
-		assert.Equal(t, resultString, "AAAAAAAAAAD////4AAAAAA==")
-	}
 }
 
 func TestAccountDetail(t *testing.T) {
