+++ clients/horizonclient/offer_request_test.go
@@ -2,9 +2,7 @@ package horizonclient
 
 import (
 	"context"
-	"fmt"
 	"testing"
-	"time"
 
 	hProtocol "github.com/stellar/go/protocols/horizon"
 	"github.com/stellar/go/support/http/httptest"
@@ -27,95 +25,6 @@ func TestOfferRequestBuildUrl(t *testing.T) {
 	// It should return valid offers endpoint and no errors
 	require.NoError(t, err)
 	assert.Equal(t, "accounts/GCLWGQPMKXQSPF776IU33AH4PZNOOWNAWGGKVTBQMIC5IMKUNP3E6NVU/offers?cursor=now&order=desc", endpoint)
-}
-
-func ExampleClient_StreamOffers() {
-	client := DefaultTestNetClient
-	// offers for account
-	offerRequest := OfferRequest{ForAccount: "GAQHWQYBBW272OOXNQMMLCA5WY2XAZPODGB7Q3S5OKKIXVESKO55ZQ7C", Cursor: "1"}
-
-	ctx, cancel := context.WithCancel(context.Background())
-	go func() {
-		// Stop streaming after 60 seconds.
-		time.Sleep(60 * time.Second)
-		cancel()
-	}()
-
-	printHandler := func(offer hProtocol.Offer) {
-		fmt.Println(offer)
-	}
-	err := client.StreamOffers(ctx, offerRequest, printHandler)
-	if err != nil {
-		fmt.Println(err)
-	}
-}
-
-func ExampleClient_NextOffersPage() {
-	client := DefaultPublicNetClient
-	// all offers
-	offerRequest := OfferRequest{ForAccount: "GAQHWQYBBW272OOXNQMMLCA5WY2XAZPODGB7Q3S5OKKIXVESKO55ZQ7C", Limit: 20}
-	offers, err := client.Offers(offerRequest)
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-	fmt.Print(offers)
-
-	// get next pages.
-	recordsFound := false
-	if len(offers.Embedded.Records) > 0 {
-		recordsFound = true
-	}
-	page := offers
-	// get the next page of records if recordsFound is true
-	for recordsFound {
-		// next page
-		nextPage, err := client.NextOffersPage(page)
-		if err != nil {
-			fmt.Println(err)
-			return
-		}
-
-		page = nextPage
-		if len(nextPage.Embedded.Records) == 0 {
-			recordsFound = false
-		}
-		fmt.Println(nextPage)
-	}
-}
-
-func ExampleClient_PrevOffersPage() {
-	client := DefaultPublicNetClient
-	// all offers
-	offerRequest := OfferRequest{ForAccount: "GAQHWQYBBW272OOXNQMMLCA5WY2XAZPODGB7Q3S5OKKIXVESKO55ZQ7C", Limit: 20}
-	offers, err := client.Offers(offerRequest)
-	if err != nil {
-		fmt.Println(err)
-		return
-	}
-	fmt.Print(offers)
-
-	// get prev pages.
-	recordsFound := false
-	if len(offers.Embedded.Records) > 0 {
-		recordsFound = true
-	}
-	page := offers
-	// get the prev page of records if recordsFound is true
-	for recordsFound {
-		// prev page
-		prevPage, err := client.PrevOffersPage(page)
-		if err != nil {
-			fmt.Println(err)
-			return
-		}
-
-		page = prevPage
-		if len(prevPage.Embedded.Records) == 0 {
-			recordsFound = false
-		}
-		fmt.Println(prevPage)
-	}
 }
 
 func TestNextOffersPage(t *testing.T) {
